data_analyst/warehouse/requirements.txt
psycopg2-binary
python-dotenv
matplotlib
jupyter
jupytext
seaborn
pandas